WIP desktop serve hot reload
